modify profile

+++ controllers/users.js
@@ -3,40 +3,43 @@ const fs = require("fs");
 const jwt = require("jsonwebtoken");
 
 const User = require("../models/user");
+const Tool = require("../models/tool");
 const userRouter = require("express").Router();
+
 const middleware = require("../utils/middleware");
 const upload = require("../utils/image-upload");
-const Tool = require("../models/tool");
 const sendVerif = require("../utils/email-verification");
 
 userRouter.post("/", async (req, res, next) => {
   try {
     const data = req.body;
-    //checking password length
+    //checking role
     if (data.role === "admin") {
       return res.status(401).end();
     }
-    if (!data.password || data.password.length < 7) {
+    //password length and verification
+    if (!data.password || data.password.length < 6) {
       return res.status(400).send("password is too short");
     } else if (data.password !== data.passwordVerification) {
       return res
         .status(400)
         .send("password and password verification dont match");
     }
-    //password hashing with bycript / await asynchrone
+    //hashing passwod with bcrypt using 10 salt rounds
     const passwordHash = await bcrypt.hash(data.password, 10);
+
     const user = new User({
       email: data.email,
       username: data.username,
       role: data.role,
       phoneNumber: data.phoneNumber,
-      firstName: data.firstName || "",
-      lastName: data.lastName || "",
+      address: data.address,
       passwordHash,
     });
 
     const createdUser = await user.save();
     res.status(201).json(createdUser);
+    //sending verification email
     await sendVerif(createdUser._id, createdUser.email);
   } catch (exception) {
     next(exception);
@@ -86,7 +89,7 @@ userRouter.put("/cart", middleware.userExtractor, async (req, res, next) => {
       await user.save();
       res.json("added to cart");
     }
-    res.end();
+    res.status(200).end();
   } catch (exception) {
     next(exception);
   }
@@ -105,19 +108,40 @@ userRouter.put(
       const imgUrl = req.file ? `/uploads/${req.file.filename}` : user.imageUrl;
 
       if (req.file && user.imageUrl !== "/uploads/avatar.jpg") {
-        fs.unlinkSync(`build${user.imageUrl}`);
+        try {
+          fs.unlinkSync(`build${user.imageUrl}`);
+        } catch (error) {
+          console.log(error);
+        }
       }
 
+      user.email = data.email || user.email;
       user.phoneNumber = data.phoneNumber || user.phoneNumber;
       user.imageUrl = imgUrl;
-      user.firstName = data.firstName || user.firstName;
-      user.lastName = data.lastName || user.lastName;
+
+      user.address.state = data.state || user.address.state;
+      user.address.city = data.city || user.address.city;
+      user.address.street = data.street || user.address.street;
+
+      user.role === "client" &&
+        (user.birthday = data.birthday || user.birthday);
+
+      if (
+        data.password &&
+        data.password.length > 6 &&
+        data.password === data.passwordVerification
+      ) {
+        const passwordHash = await bcrypt.hash(data.password, 10);
+        user.passwordHash = passwordHash;
+      }
 
       const updatedUser = await user.save();
       res.json(updatedUser);
     } catch (exception) {
-      fs.unlinkSync(req.file.path);
       next(exception);
+      if (req.file) {
+        fs.unlinkSync(req.file.path);
+      }
     }
   }
 );
